Refining dashboard and add transactions

+++ every-rupee/src/main/java/in/co/everyrupee/controller/income/UserTransactionsController.java
@@ -1,12 +1,11 @@
 package in.co.everyrupee.controller.income;
 
+import java.security.Principal;
 import java.util.List;
 
-import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -14,10 +13,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
-import in.co.everyrupee.exception.ResourceNotFoundException;
 import in.co.everyrupee.pojo.income.UserTransaction;
 import in.co.everyrupee.repository.income.UserTransactionsRepository;
-import in.co.everyrupee.security.core.userdetails.MyUser;
 import in.co.everyrupee.service.income.IUserTransactionService;
 
 /**
@@ -40,18 +37,20 @@ public class UserTransactionsController {
 
     // Get a Single User Transaction
     @RequestMapping(value = "/", method = RequestMethod.GET)
-    public List<UserTransaction> getUserTransactionByUserId() {
-	MyUser user = (MyUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-	List<UserTransaction> userTransactions = userTransactionsRepository.findByUserId(user.getId());
-	if (CollectionUtils.isEmpty(userTransactions)) {
-	    throw new ResourceNotFoundException("UserTransactions", "userId", user.getId());
+    public List<UserTransaction> getUserTransactionByUserId(Principal userPrincipal) {
+	if (userPrincipal == null) {
+	    throw new SecurityException();
 	}
-	return userTransactions;
+
+	return userTransactionService.fetchUserTransaction();
     }
 
     // Update a UserTransaction
     @RequestMapping(value = "/save", method = RequestMethod.POST, consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-    public UserTransaction update(@RequestBody MultiValueMap<String, String> formData) {
+    public UserTransaction update(@RequestBody MultiValueMap<String, String> formData, Principal userPrincipal) {
+	if (userPrincipal == null) {
+	    throw new SecurityException();
+	}
 
 	UserTransaction userTransactionResponse = userTransactionService.saveUserTransaction(formData);
 	return userTransactionResponse;
@@ -59,7 +58,10 @@ public class UserTransactionsController {
 
     // Delete a User Transaction
     @RequestMapping(value = "/{transactionIds}", method = RequestMethod.DELETE)
-    public ResponseEntity<?> deleteUserTransactionById(@PathVariable String transactionIds) {
+    public ResponseEntity<?> deleteUserTransactionById(@PathVariable String transactionIds, Principal userPrincipal) {
+	if (userPrincipal == null) {
+	    throw new SecurityException();
+	}
 
 	userTransactionService.deleteUserTransactions(transactionIds);
